Delivery price implemented and test passed.

+++ src/main/java/strategy_pattern/DeliveryMethod.java
@@ -1,5 +1,5 @@
 package strategy_pattern;
 
 public interface DeliveryMethod {
-    int deliveryPrice(int weight);
+    double deliveryPrice(double weight);
 }
